DocumentAnalyzerProject/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
nltk==3.5
gunicorn==20.0.4